refactor: move implementation out of index file
